Player damage to monsters and other players now done

+++ src/com/herocraftonline/dev/heroes/damage/HeroesDamage.java
@@ -6,20 +6,20 @@ import com.herocraftonline.dev.heroes.Heroes;
 
 public class HeroesDamage {
     public Heroes plugin;
-    private HashMap<Integer, Integer> mobHealthValues = new HashMap<Integer, Integer>();
+    private HashMap<Integer, Double> mobHealthValues = new HashMap<Integer, Double>();
 
     public HeroesDamage(Heroes plugin) {
         this.plugin = plugin;
     }
 
-    public void addMonster(Integer entityID, Integer hp) {
+    public void addMonster(Integer entityID, Double hp) {
         mobHealthValues.put(entityID, hp);
     }
 
     /**
      * @return the mobHealthValues
      */
-    public HashMap<Integer, Integer> getMobHealthValues() {
+    public HashMap<Integer, Double> getMobHealthValues() {
         return mobHealthValues;
     }
